Handle classes with types and duplicated right hand sides.
The param parsing code was a little naive before and didn't
take in to account that you can have the same value on the
right hand side of an assignment. This commit adds specs to
test for this and code to make it work.

Also add an example that uses types to ensure I don't cause a
regression.